requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Auth / Security
python-jose[cryptography]>=3.3.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
rich>=13.0.0       # For beautiful terminal output
httpx>=0.25.0      # For async HTTP requests
